dcim: rearrange models, fixing #3092
