+++ netbox/dcim/models/__init__.py
@@ -1,0 +1,3 @@
+from .models import *
+from .component_models import *
+from .component_template_models import *
